add (unused) taiko hit position circle
Overloaded approach circle in taiko directory, which does not work.

Since std is skinned, it would be a BAD idea to replace the root-directory
approach circle, as to the eyes of an std player it does none other than
adding clutter.

Leaving the file here for reference. Won't make SD version.